resolves bug #225 - trunc int64 to int; no large file support anymore

+++ src/DbFp.fs
@@ -39,7 +39,7 @@ type DbFpBlock = {
 
 type DbFpDat = {
              id : Key128.t;
-             len : int;
+             len : int64;
              osusr : string;
              osgrp : string;
              osattr : string;
@@ -52,7 +52,7 @@ type DbFp() =
     inherit DbCtrl<string, DbFpDat>()
 
     let emptyFp : DbFpDat = { id = Key128.create();
-                              len=0; osusr=""; osgrp=""; osattr="";
+                              len=0L; osusr=""; osgrp=""; osattr="";
                               checksum = Key256.fromHex "deadbeefcafecafedeadbeefcafecafedeadbeefcafecafedeadbeefcafecafe"; blocks=[] }
 
     let rec inBlock (reader : XmlTextReader) (l : DbFpBlock list) =
@@ -102,7 +102,7 @@ type DbFp() =
             else record
         elif reader.Name = "length" && reader.NodeType = Xml.XmlNodeType.Element then
             if reader.Read() && reader.NodeType = Xml.XmlNodeType.Text then
-                inAttrs {record with len = Int32.Parse(reader.Value)} reader
+                inAttrs {record with len = Int64.Parse(reader.Value)} reader
             else record
         elif reader.Name = "last" && reader.NodeType = Xml.XmlNodeType.Element then
             if reader.Read() && reader.NodeType = Xml.XmlNodeType.Text then
